[M] Move mirrors to its own package
