+++ src/main/java/org/hydev/mcpm/client/arguments/CommandsFactory.java
@@ -4,8 +4,7 @@ import org.hydev.mcpm.client.arguments.parsers.*;
 import org.hydev.mcpm.client.commands.entries.*;
 import org.hydev.mcpm.client.database.ListAllInteractor;
 import org.hydev.mcpm.client.database.LocalPluginTracker;
-import org.hydev.mcpm.client.database.Mirror;
-import org.hydev.mcpm.client.database.MirrorSelector;
+import org.hydev.mcpm.client.database.mirrors.MirrorSelector;
 import org.hydev.mcpm.client.database.export.ExportInteractor;
 import org.hydev.mcpm.client.database.fetcher.LocalDatabaseFetcher;
 import org.hydev.mcpm.client.database.searchusecase.SearchInteractor;
